refactor. ResultView에서만 쓰이는 외부 유틸 converToStringFromList()를 ResultView 내부 메소드로 리팩토링
