--- src/main/java/step3to5/utils/ConvertToStringFromList.java
@@ -1,13 +1,0 @@
-package step3to5.utils;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class ConvertToStringFromList {
-    public static String convertToStringFromList(List<String> list) {
-        String result = list.stream()
-                .map(String::valueOf)
-                .collect(Collectors.joining(", "));
-        return result;
-    }
-}
